requirements.txt
antiorm==1.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
db==0.1.1
db-sqlite3==0.0.1
et-xmlfile==1.1.0
idna==3.4
numpy==1.24.1
openai==0.25.0
openpyxl==3.0.10
pandas==1.5.2
pandas-stubs==1.5.2.230105
pyTelegramBotAPI==4.9.0
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
requests==2.28.1
six==1.16.0
tqdm==4.64.1
types-pytz==2022.7.0.0
typing-extensions==4.4.0
urllib3==1.26.13
